requirements.txt
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==2.1.1
contourpy==1.2.0
cycler==0.12.1
einops==0.7.0
filelock==3.9.0
fonttools==4.48.1
fsspec==2023.4.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kaggle==1.6.8
kiwisolver==1.4.5
lightning-utilities==0.10.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
normflows==1.7.3
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pygame==2.5.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2024.1
requests==2.28.1
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.4.0
torchmetrics==1.3.0.post0
tqdm==4.66.2
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
webencodings==0.5.1
